+++ terra-rust-api/src/client/core_types.rs
@@ -32,7 +32,7 @@ impl Coin {
         }
     }
     /// Parse the string "nnnnnXXXX" format where XXXX is the coin type
-    pub fn parse(str: &str) -> anyhow::Result<Option<Coin>> {
+    pub fn parse(str: &str) -> Result<Option<Coin>, TerraRustAPIError> {
         lazy_static! {
             static ref RE: Regex = Regex::new(r"^(\d+[.]?\d*)([a-zA-Z]+)$").unwrap();
             static ref RE_IBC: Regex = Regex::new(r"^(\d+[.]?\d*)(ibc/[a-fA-F0-9]+)$").unwrap();
@@ -56,21 +56,20 @@ impl Coin {
     /// this will take a comma delimited string of coins and return a sorted (by denom) vector of coins
     /// eg "22.707524482460197756uaud,21.882510617180501989ucad,16.107413560222631626uchf,114.382279464849248732ucny,14.594888140543189388ueur,12.689498975492463452ugbp,136.932658449160933002uhkd,1315.661396873891976912uinr,1917.803659404458501345ujpy,20710.846165266109229516ukrw,50292.255931832196576203umnt,12.276992042852615569usdr,23.395036036859944228usgd,0.0uthb,17.639582167170638049uusd"
     ///
-    pub fn parse_coins(str: &str) -> anyhow::Result<Vec<Coin>> {
+    pub fn parse_coins(str: &str) -> Result<Vec<Coin>, TerraRustAPIError> {
         let vec_res_opt_coins = str
             .split(',')
             .map(Coin::parse)
-            .collect::<Vec<anyhow::Result<Option<Coin>>>>();
+            .collect::<Vec<Result<Option<Coin>, TerraRustAPIError>>>();
         let mut coins: Vec<Coin> = Vec::with_capacity(vec_res_opt_coins.len());
         for vroc in vec_res_opt_coins {
-            let coin_opt = vroc.map_err(|source| TerraRustAPIError::CoinParseErrV {
+            let coin_opt = vroc.map_err(|_source| TerraRustAPIError::CoinParseErrV {
                 parse: str.parse().unwrap(),
-                source,
             })?;
 
             match coin_opt {
                 None => {
-                    return Err(TerraRustAPIError::CoinParseErr(str.parse().unwrap()).into());
+                    return Err(TerraRustAPIError::CoinParseErr(str.parse().unwrap()));
                 }
                 Some(coin) => {
                     coins.push(coin);