+++ terra-rust-api/src/client/fcd.rs
@@ -1,3 +1,4 @@
+use crate::errors::TerraRustAPIError;
 use crate::Terra;
 use rust_decimal::Decimal;
 use std::collections::HashMap;
@@ -10,7 +11,7 @@ impl FCD<'_> {
     pub fn create<'a>(terra: &'a Terra, fcd_url: &'a str) -> FCD<'a> {
         FCD { terra, fcd_url }
     }
-    pub async fn gas_prices(&self) -> anyhow::Result<HashMap<String, Decimal>> {
+    pub async fn gas_prices(&self) -> Result<HashMap<String, Decimal>, TerraRustAPIError> {
         Ok(self
             .terra
             .send_cmd_url::<HashMap<String, Decimal>>(self.fcd_url, "/v1/txs/gas_prices", None)
@@ -19,7 +20,7 @@ impl FCD<'_> {
     pub async fn fetch_gas_prices(
         client: &reqwest::Client,
         fcd_url: &str,
-    ) -> anyhow::Result<HashMap<String, Decimal>> {
+    ) -> Result<HashMap<String, Decimal>, TerraRustAPIError> {
         Ok(Terra::fetch_url::<HashMap<String, Decimal>>(
             client,
             fcd_url,