+++ terra-rust-api/Changelog.md
@@ -3,7 +3,8 @@
 ## 1.2
 ### 1.2.4 - Devel
 * Serialize some structures
-* new fn to get delegators for a validator with a limit 
+* new fn to get delegators for a validator with a limit
+* Error messages converted to use TerraRustAPIError
 ### 1.2.3 - 14-Feb-22
 * PR#8 : use default for sequence deserialize thanks @patternmachine!
 ### 1.2.2 - 6-Feb-22